# packages/backend/src/OpenApi/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API Documentation",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/API",
      "description": "Base URL for the API"
    }
  ],
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "id_token",
        "description": "This uses authenticaion token generated with a secret obtained from keycloak, the token can be sent both as a cookie named id_token or in the auth_key property of the election object. [More information about token generation can be found here](https://github.com/Equal-Vote/star-server/blob/6fdb2f653266b5f2710f0701509f54e1e558ecc1/docs/api.md)<br><br>For API testing [log in here](https://keycloak.prod.equal.vote/realms/Prod/protocol/openid-connect/auth?client_id=web&response_type=code&redirect_uri=https://bettervoting.com/API/Docs&scope=openid)"
      }
    },
    "schemas": {
...
      "ElectionResults": {
        "anyOf": [
          {
            "$ref": "#/components/schemas/starResults"
          },
          {
            "$ref": "#/components/schemas/allocatedScoreResults"
          },
          {
            "$ref": "#/components/schemas/approvalResults"
          },
          {
            "$ref": "#/components/schemas/rankedRobinResults"
          },
          {
            "$ref": "#/components/schemas/irvResults"
          },
          {
            "$ref": "#/components/schemas/pluralityResults"
          }
        ]
      },
...
      "roles": {
        "enum": [
          "owner",
          "admin",
          "auditor",
          "system_admin",
          "credentialer"
        ],
        "type": "string"
      },
...
    }
  },
  "paths": {
...
    "/Election/{id}": {
      "get": {
        "summary": "Get election by ID",
        "tags": [
          "Elections"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "The election ID"
          }
        ],
        "responses": {
          "200": {
            "description": "The election description by ID",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "$ref": "#/components/schemas/Election"
                }
              }
            }
          },
          "404": {
            "description": "Election not found"
          }
        }
      },
      "delete": {
        "summary": "Delete election by ID",
        "tags": [
          "Elections"
        ],
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "The election ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Election deleted"
          },
          "404": {
            "description": "Election not found"
          }
        }
      }
    },
    "/Election/{_id}/exists": {
      "get": {
        "summary": "Check if election exists by ID",
        "tags": [
          "Elections"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "The election ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Election exists",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "exists": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Election not found"
          }
        }
      }
    },
    "/Election/{id}/claim": {
      "get": {
        "summary": "Claims a election that was created without an account",
        "tags": [
          "Elections"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "The election ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "claim_key": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "If claim was successful"
          },
          "404": {
            "description": "Election not found"
          }
        }
      }
    },
    "/Elections": {
      "get": {
        "summary": "Get all elections",
        "tags": [
          "Elections"
        ],
        "responses": {
          "200": {
            "description": "List of elections",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "elections_as_official": {
                      "oneOf": [
                        {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/Election"
                          }
                        },
                        {
                          "type": null
                        }
                      ]
                    },
                    "elections_as_unsubmitted_voter": {
                      "oneOf": [
                        {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/Election"
                          }
                        },
                        {
                          "type": null
                        },
                        {
                          "type": "undefined"
                        }
                      ]
                    },
                    "elections_as_submitted_voter": {
                      "oneOf": [
                        {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/Election"
                          }
                        },
                        {
                          "type": null
                        }
                      ]
                    },
                    "open_elections": {
                      "oneOf": [
                        {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/Election"
                          }
                        },
                        {
                          "type": null
                        }
                      ]
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new election",
        "tags": [
          "Elections"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "Election": {
                    "$ref": "#/components/schemas/NewElection"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Election created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "election": {
                      "type": "object",
                      "$ref": "#/components/schemas/Election"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/QueryElections": {
      "get": {
        "summary": "Query Elections based on time range (sys-admin only)",
        "tags": [
          "Elections"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "start_time": {
                    "type": "date"
                  },
                  "end_time": {
                    "type": "date"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "List of elections",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "open_elections": {
                      "oneOf": [
                        {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/Election"
                          }
                        },
                        {
                          "type": null
                        }
                      ]
                    },
                    "closed_elections": {
                      "oneOf": [
                        {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/Election"
                          }
                        },
                        {
                          "type": null
                        },
                        {
                          "type": "undefined"
                        }
                      ]
                    },
                    "popular_elections": {
                      "oneOf": [
                        {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/Election"
                          }
                        },
                        {
                          "type": null
                        }
                      ]
                    },
                    "vote_counts": {
                      "oneOf": [
                        {
                          "type": "array",
                          "items": {
                            "type": "object"
                          }
                        },
                        {
                          "type": null
                        }
                      ]
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/Election/{id}/edit": {
      "post": {
        "summary": "Edit an election",
        "tags": [
          "Elections"
        ],
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "The election ID"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "Election": {
                    "$ref": "#/components/schemas/Election"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Election edited",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "election": {
                      "type": "object",
                      "$ref": "#/components/schemas/Election"
                    },
                    "voterAuth": {
                      "type": "object",
                      "properties": {
                        "authorized_voter": {
                          "type": "boolean"
                        },
                        "has_voted": {
                          "type": "boolean"
                        },
                        "roles": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/roles"
                          }
                        },
                        "permissions": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/permissions"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Election not found"
          }
        }
      }
    },
...
    "/ElectionResult/{id}": {
      "get": {
        "summary": "Get election results by ID",
        "tags": [
          "Elections"
        ],
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "The election ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Election results",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "election": {
                      "type": "object",
                      "$ref": "#/components/schemas/Election"
                    },
                    "results": {
                      "type": "object",
                      "$ref": "#/components/schemas/ElectionResults"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Election not found"
          }
        }
      }
    },
...
    "/Election/{id}/sendInvites": {
      "post": {
        "summary": "Send invitations for an election",
        "tags": [
          "Elections"
        ],
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "The election ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Invitations sent"
          },
          "404": {
            "description": "Election not found"
          }
        }
      }
    },
...
  },
...
}